build: use -Wextra to harden config

+++ main_board/app/src/optics/mirrors/mirrors.c
@@ -415,10 +415,10 @@ mirrors_angle_relative(int32_t angle_millidegrees, mirror_t mirror)
         (int32_t)roundf(sinf((float)angle_millidegrees * M_PI / 360000.0f) *
                         motors_arm_length[mirror] * (float)steps_per_mm);
     int32_t xtarget = x + steps;
-    if (xtarget >
-            (motors_refs[mirror].x0 + motors_refs[mirror].full_course / 2) ||
-        xtarget <
-            (motors_refs[mirror].x0 - motors_refs[mirror].full_course / 2)) {
+    if (xtarget > (int32_t)(motors_refs[mirror].x0 +
+                            motors_refs[mirror].full_course / 2) ||
+        xtarget < (int32_t)(motors_refs[mirror].x0 -
+                            motors_refs[mirror].full_course / 2)) {
         return RET_ERROR_FORBIDDEN;
     }
